requirements.txt
aniso8601==7.0.0
APScheduler==3.7.0
blinker==1.4
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
croniter==1.0.10
dnspython==2.1.0
dynaconf==3.1.4
email-validator==1.1.2
firebase-admin==4.5.3
Flask==1.1.2
Flask-APScheduler==1.12.1
Flask-Cache==0.13.1
Flask-Caching==1.10.1
Flask-Compressor==0.2.0
Flask-Cors==3.0.10
Flask-GraphQL==2.0.1
Flask-GraphQL-Auth==1.3.2
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
graphene==2.1.8
graphene-mongo==0.2.13
graphene-sqlalchemy==2.3.0
graphene-sqlalchemy-filter==1.12.2
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==1.0.0
grpcio==1.36.1
httplib2==0.19.1
idna==2.10
iso8601==0.1.14
itsdangerous==1.1.0
Jinja2==2.11.3
kea-cerberus==1.0
MarkupSafe==1.1.1
mongoengine==0.23.0
mongomock==3.22.1
msgpack==1.0.2
natsort==7.1.1
numpy
packaging==20.9
phonenumbers==8.12.20
Pillow==8.2.0
promise==2.3
proto-plus==1.18.1
protobuf==3.15.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.0
pymongo==3.11.3
PyMySQL==1.0.2
pyparsing==2.4.7
python-box==5.3.0
python-dateutil==2.8.1
python-dotenv==0.17.0
python-slugify==4.0.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
Rx==1.6.1
schedule==1.0.0
sentinels==1.0.0
singledispatch==3.6.1
six==1.15.0
SQLAlchemy==1.4.5
text-unidecode==1.3
toml==0.10.2
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
uWSGI==2.0.19.1
Werkzeug==1.0.1
WTForms==2.3.3